Util Changes, and update to README
Util now includes mountNamespace to load an API into a Runtime environment for direct manipulation.
This is consistent with the "conqur" view of the World that processes are the agent's of change but APIs are the interface from the programmer to those agents.
This is bringing conqur inline with the API/Process patterns of Elixir.
Added Utility functions for reduce, filter, and other conveniences.  Most are copied shamelessly and very directly from https://github.com/nanoutils/nanoutils which is under the MIT License